Trait proofs::Verifier

pub trait Verifier: Hasher {
    // Required methods
    fn hash_of(a: Self::Hash, b: Self::Hash) -> Self::Hash;
    fn initial_matches(&self, doc_root: Self::Hash) -> Option<Vec<Self::Hash>>;

    // Provided methods
    fn verify_proofs(
        &self,
        doc_root: Self::Hash,
        proofs: &[Proof<Self::Hash>]
    ) -> bool { ... }
    fn verify_proof(
        &self,
        doc_root: Self::Hash,
        proof: &Proof<Self::Hash>
    ) -> bool { ... }
}

Required Methods

fn hash_of(a: Self::Hash, b: Self::Hash) -> Self::Hash

Computes hash of the a + b using hash function

fn initial_matches(&self, doc_root: Self::Hash) -> Option<Vec<Self::Hash>>

Returns the initial set of hashes to verify proofs. None implies a failed proof verification

Provided Methods

fn verify_proofs( &self, doc_root: Self::Hash, proofs: &[Proof<Self::Hash>] ) -> bool

Verifies each proof and return true if all the proofs are valid else returns false

fn verify_proof(&self, doc_root: Self::Hash, proof: &Proof<Self::Hash>) -> bool

Verifies the proof and returns true if valid